Refactor htop cleanup and enhance task display with load information

+++ scheduler/sched_info.h
@@ -9,7 +9,7 @@
 typedef struct sched_info {
     task_state_t state;
 
-    uint64_t load;
+    uint32_t load;
     double load_contrib;
     
     uint32_t quantum;
